Improved logging.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/filevector@311 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ tests/TestUtil.h
@@ -35,8 +35,8 @@ public :
     }
 
     static void create_and_fill_variable(unsigned int  nobs, float * in );
-    static void create_empty_filevector(string tmp_file_name,unsigned int  nvars, unsigned int  nobs);
-    static void create_empty_filevector(string tmp_file_name ,unsigned int  nvars, unsigned int  nobs, unsigned short int type);
+    static void create_empty_filevector(string tmp_file_name,unsigned long nvars, unsigned long nobs);
+    static void create_empty_filevector(string tmp_file_name,unsigned long nvars, unsigned long nobs, unsigned short int type);
     static bool compare_arrays(float * a1,float * a2, int size);
 };
